==> 3_failure_causes.ipynb <==
# 3. Failure causes
Identify most common causes of failure for these trials using the “reason_for_stop” field in the CTRP API.

- Lack of accrual
- Lack of efficacy
- Toxicity
- Sponsor withdrawal
- Others?


from collections import defaultdict, Counter
import pandas as pd
import os, glob
import requests
import json
import re
from scipy import stats


%%time
active_trials = json.load(open('interventional_trials_with_descendants2024-07-26.json','r'))


len(active_trials['data']), active_trials['total']


biomarker_trials = []
nonbiomarker_trials = []

for study in active_trials['data']:
    genes = [g for g in study['biomarkers_new']['inclusion']['TREE']['symbols_dz'] if type(g)==str]
    if len(genes) >0:
        biomarker_trials.append(study)
    else:
        nonbiomarker_trials.append(study)
  

trial_data = active_trials['data']

keys_arr = sorted(trial_data[0].keys())
keys_to_values = {}
for key in keys_arr:
    # print('******')
    # print(key)
    key_list = []
    for trial in trial_data:
        key_list.append(trial[key])
    try:
        key_list_counter = Counter(key_list)
        keys_to_values[key] = key_list_counter
        # if len(key_list_counter)<20:
        #     print(key_list_counter)
    except:
        print('ERROR: not typable list',key)

studystop_dict = keys_to_values['why_study_stopped']
len(studystop_dict)

studystop_dict[None], 

sum(studystop_dict.values())

sum(studystop_dict.values())-studystop_dict[None]

1-studystop_dict[None]/sum(studystop_dict.values())

clean up the 2334 reasons studies stopped
17.7% of studies failed or 17196 of the 20894 interventional trials

studystop_dict_filt = defaultdict(int)
for key, value in studystop_dict.items():
    studystop_dict_filt[str(key).lower()]+=value


len(studystop_dict_filt)

studystop_dict_filt_df = pd.DataFrame({'reason':studystop_dict_filt.keys(), 'num_trials':studystop_dict_filt.values()}).sort_values(by=['num_trials'],ascending=False)
display(studystop_dict_filt_df[:5])

# studystop_dict_filt_df.to_csv('studystop_reason_df.csv')

manual annotation

# then divide by phase

studystop_dict_filt_df_annon  = pd.read_excel('studystop_reason_df.xlsx')
studystop_dict_filt_df_annon['reason'] = studystop_dict_filt_df_annon['reason'].str.replace("_x000D_", "\r")
studystop_dict_filt_df_annon['reason'] = studystop_dict_filt_df_annon['reason'].str.replace('\W', '')
studystop_dict_filt_df_annon.reason.replace({r'[^\x00-\x7F]+':''}, regex=True, inplace=True)
studystop_dict_filt_df_annon.reason.replace({r'[^\w\s]':''}, regex=True, inplace=True)


reason_to_group = pd.Series(studystop_dict_filt_df_annon.group.values, index=studystop_dict_filt_df_annon.reason.values).to_dict()


def get_biomarker_status(study):
    genes = [g for g in study['biomarkers_new']['inclusion']['TREE']['symbols_dz'] if type(g)==str]
    if len(genes) >0:
        return 'biomarker'
    else:
        return 'nonbiomarker'

def get_phase_study(study):
    """
    get phase of trial as a string, if multi-phase trial, trial gets upgraded to highest phase
    input:
    study <dict>
    return <str> phase
    """
    phase_str = str(study['phase']).strip()
    if phase_str=='I':
        return 'I'
    elif phase_str=='II' or phase_str =='I_II':
        return 'II'
    elif phase_str=='III' or phase_str =='II_III':
        return 'III'
    elif phase_str=='IV':
        return 'IV'
    elif phase_str=='O':
        return 'other'
    elif phase_str=='NA':
        return 'other'
    else:
        print(phase_str, type(phase_str), phase_str=='0' )
        return 'other'
    

phase_to_reason_count = defaultdict(list)
phase_to_reason_count_biomarker = defaultdict(list)
phase_to_reason_count_nonbiomarker = defaultdict(list)


trial_id_to_phasegene_info = {}
for trial in active_trials['data']:
    phase = get_phase_study(trial)
    biomarker_status = get_biomarker_status(trial)
    genes = [g for g in trial['biomarkers_new']['inclusion']['TREE']['symbols_dz'] if type(g)==str]
    str_reason = trial['why_study_stopped']
    if str_reason is not None:
        str_reason = str_reason.lower()
        str_reason = re.sub(r'[^\w\s]','',str_reason)
        str_reason = re.sub(r'[^\x00-\x7F]','',str_reason)
        

        if str_reason not in reason_to_group:
            group = 'other'
        else:
            group = reason_to_group[str_reason]
        phase_to_reason_count[phase].append(group)
        phase_to_reason_count['total'].append(group)
        if biomarker_status =='biomarker':
            phase_to_reason_count_biomarker[phase].append(group)
            phase_to_reason_count_biomarker['total'].append(group)
        else:
            phase_to_reason_count_nonbiomarker[phase].append(group)
            phase_to_reason_count_nonbiomarker['total'].append(group)
            
    
        trial_id_to_phasegene_info[trial['nci_id']] = {'failure':group,'phase':phase,'biomarker_status':biomarker_status,'genes_arr':genes}

def process_and_save_files(phase_to_reason_dict, prefix='phase_to_reason_df'):
    phase_to_reason_counter = {}
    for phase, reason_list in phase_to_reason_dict.items():
        phase_to_reason_counter[phase] = Counter(reason_list)
        
    phase_to_reason_df = pd.DataFrame(phase_to_reason_counter).fillna(0)
    phase_to_reason_df.to_csv(prefix+'.csv')
    phase_to_reason_df_total = phase_to_reason_df.total
    phase_to_reason_df = phase_to_reason_df[['I', 'II','III','IV','other']]
    nl = "\n"
    chi2, p, dof, expected = stats.chi2_contingency(phase_to_reason_df)
    print(f"Chi2 value= {chi2}{nl}p-value= {p}{nl}Degrees of freedom= {dof}{nl}")
    phase_to_reason_df['I'] = phase_to_reason_df['I']/phase_to_reason_df_total/sum(phase_to_reason_df['I'])*10000
    phase_to_reason_df['II'] = phase_to_reason_df['II']/phase_to_reason_df_total/sum(phase_to_reason_df['II'])*10000
    phase_to_reason_df['III'] = phase_to_reason_df['III']/phase_to_reason_df_total/sum(phase_to_reason_df['III'])*10000
    phase_to_reason_df['IV'] = phase_to_reason_df['IV']/phase_to_reason_df_total/sum(phase_to_reason_df['IV'])*10000
    phase_to_reason_df['other'] = phase_to_reason_df['other']/phase_to_reason_df_total/sum(phase_to_reason_df['other'])*10000
    phase_to_reason_df.to_csv(prefix+'_norm.csv')
    
    # display(phase_to_reason_df)

    
    phase_to_reason_df["reason"] = phase_to_reason_df.index
    phase_to_reason_df_long = phase_to_reason_df.melt('reason')
    phase_to_reason_df_long.columns = ['reason', 'phase','norm']
    phase_to_reason_df_long.to_csv(prefix+'_long.csv')
    
    display(phase_to_reason_df_total/sum(phase_to_reason_df_total))

    return phase_to_reason_df_total

failure_all = process_and_save_files(phase_to_reason_count, prefix='phase_to_reason_df')

failure_biomarker = process_and_save_files(phase_to_reason_count_biomarker, prefix='phase_to_reason_df_biomarker')

failure_nonbiomarker = process_and_save_files(phase_to_reason_count_nonbiomarker, prefix='phase_to_reason_df_nonbiomarker')


comp_failure = pd.DataFrame({'biomarker':failure_biomarker,'nonbiomarker':failure_nonbiomarker})
chi2, p, dof, expected = stats.chi2_contingency(comp_failure)
nl = "\n"
print(f"Chi2 value= {chi2}{nl}p-value= {p}{nl}Degrees of freedom= {dof}{nl}")

print('table after adding all which is sum of biomarker and nonbiomarker')
comp_failure['all'] = comp_failure.sum(axis=1)
comp_failure/comp_failure.sum(axis=0)

biomarker trials and nonbiomarker trials have have different failure causes (Chi squared test, p value = 5.8e-3). Interestingly, while overall, accrual was the most likely cause of trial stoppage at 28.7%, problems with sponsorship was the most common cause of failure for biomarker trials (30.3%) and a higher percentage compared to that for nonbiomarker trials (22.9%). 

# find genes go term for phase 1 failure reasons
hypothesis; there are pathways we shouldn't target in early phases 


trial_id_to_phasegene_info_df = pd.DataFrame.from_dict(trial_id_to_phasegene_info,orient='index')

trial_id_to_phasegene_info_df.to_csv('trial_id_to_phasegene_info_df.csv')

trial_id_to_phasegene_info_df[:5]

trial_id_to_phasegene_info_df.failure.value_counts()

genes_failure = []
genes_failure_phase1 = []
for idx, row in trial_id_to_phasegene_info_df.iterrows():
    genes_failure +=row['genes_arr']
    if row['phase'] =='I':
        genes_failure_phase1 +=row['genes_arr']

len(set(genes_failure)), len(set(genes_failure_phase1))

genes_failure_efficacy = []
genes_failure_phase1_efficacy = []
genes_failure_otherphase_efficacy = []

genes_failure_accrual = []
genes_failure_phase1_accrual = []
genes_failure_otherphase_accrual = []


genes_failure_safety = []
genes_failure_phase1_safety = []
genes_failure_otherphase_safety = []

for idx, row in trial_id_to_phasegene_info_df.iterrows():
    if row['failure'] =='efficacy':
    
        genes_failure_efficacy +=row['genes_arr']
        if row['phase'] =='I':
            genes_failure_phase1_efficacy +=row['genes_arr']
        elif (row['phase'] =='II') or (row['phase'] =='III') or (row['phase'] =='IV') :
            genes_failure_otherphase_efficacy +=row['genes_arr']

    elif row['failure'] =='accrual':
    
        genes_failure_accrual +=row['genes_arr']
        if row['phase'] =='I':
            genes_failure_phase1_accrual +=row['genes_arr']
        elif (row['phase'] =='II') or (row['phase'] =='III') or (row['phase'] =='IV') :
            genes_failure_otherphase_accrual +=row['genes_arr']

            
    elif row['failure'] =='safety':
    
        genes_failure_safety +=row['genes_arr']
        if row['phase'] =='I':
            genes_failure_phase1_safety +=row['genes_arr']
        elif (row['phase'] =='II') or (row['phase'] =='III') or (row['phase'] =='IV') :
            genes_failure_otherphase_safety +=row['genes_arr']
            
            
genes_failure_efficacy = set(genes_failure_efficacy)
genes_failure_phase1_efficacy = set(genes_failure_phase1_efficacy)
genes_failure_otherphase_efficacy = set(genes_failure_otherphase_efficacy)


genes_failure_accrual = set(genes_failure_accrual)
genes_failure_phase1_accrual = set(genes_failure_phase1_accrual)
genes_failure_otherphase_accrual = set(genes_failure_otherphase_accrual)


genes_failure_safety = set(genes_failure_safety)
genes_failure_phase1_safety = set(genes_failure_phase1_safety)
genes_failure_otherphase_safety = set(genes_failure_otherphase_safety)



len(genes_failure_efficacy), len(genes_failure_phase1_efficacy)

len(genes_failure_safety), len(genes_failure_phase1_safety), len(genes_failure_otherphase_safety)

genes_failure_phase1_safety

# genes_failure_otherphase_efficacy

# these are phase 1 only failure genes for efficacy
print(len(genes_failure_phase1_efficacy - genes_failure_otherphase_efficacy))
for g in sorted(genes_failure_phase1_efficacy - genes_failure_otherphase_efficacy):
    print(g)
    
sorted(genes_failure_phase1_efficacy - genes_failure_otherphase_efficacy)

# these are phase 2+ only failure genes - for efficacy
print(len(genes_failure_otherphase_efficacy - genes_failure_phase1_efficacy))
for g in sorted(genes_failure_otherphase_efficacy - genes_failure_phase1_efficacy):
    print(g)
    
sorted(genes_failure_otherphase_efficacy - genes_failure_phase1_efficacy)

#all phase genes
genes_failure_phase1_efficacy.intersection( genes_failure_otherphase_efficacy)

# these are phase 1 only failure genes for accrual
print(len(genes_failure_phase1_accrual - genes_failure_otherphase_accrual))
for g in sorted(genes_failure_phase1_accrual - genes_failure_otherphase_accrual):
    print(g)
    
sorted(genes_failure_phase1_accrual - genes_failure_otherphase_accrual)

# these are phase 2+ only failure genes - for accrual
print(len(genes_failure_otherphase_accrual - genes_failure_phase1_accrual))
for g in sorted(genes_failure_otherphase_accrual - genes_failure_phase1_accrual):
    print(g)
    
sorted(genes_failure_otherphase_accrual - genes_failure_phase1_accrual)

#all phase genes - accrual
genes_failure_phase1_accrual.intersection( genes_failure_otherphase_accrual)
